third_party/WebKit/LayoutTests/http/tests/csslayout/supports-insecure-context.html

<!DOCTYPE html>
<meta name="assert" content="This test checks that CSS.supports() correctly returns false for a layout() function on an insecure context." />
<script src="../resources/get-host-info.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>
  if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
    window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
  } else {
    test(function() {
      assert_false(CSS.supports('display', 'layout(foo)'));
    });
  }
</script>
